pyproject.toml
[tool.poetry]
name = "contest-bot"
version = "0.1.0"
description = ""
authors = ["Tblkva <53375742+Tblkva@users.noreply.github.com>"]
readme = "README.md"
packages = [{include = "app"}, {include="app.py"}]

[tool.poetry.dependencies]
python = "^3.8"
vkbottle = "^4.3.12"
loguru = "^0.7.0"
aiosqlite = "^0.19.0"
python-dotenv = "^1.0.0"
msgpack = "^1.0.5"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
